move integration tests to a separate class, to prevent running against network in travis
