+++ kin-core/src/androidTest/java/kin/core/KinAccountTest.java
@@ -1,0 +1,73 @@
+package kin.core;
+
+
+import static junit.framework.Assert.assertNull;
+
+import android.support.test.InstrumentationRegistry;
+import java.io.IOException;
+import java.math.BigDecimal;
+import kin.core.exception.AccountDeletedException;
+import org.junit.After;
+import org.junit.Before;
+import org.junit.Rule;
+import org.junit.Test;
+import org.junit.rules.ExpectedException;
+
+@SuppressWarnings({"deprecation", "ConstantConditions"})
+public class KinAccountTest {
+
+    private static final String TEST_NETWORK_URL = "https://horizon-testnet.stellar.org";
+    private KinClient kinClient;
+
+    @Rule
+    public ExpectedException expectedEx = ExpectedException.none();
+
+    @Before
+    public void setup() throws IOException {
+        kinClient = new KinClient(InstrumentationRegistry.getTargetContext(),
+            new ServiceProvider(TEST_NETWORK_URL, ServiceProvider.NETWORK_ID_TEST));
+        kinClient.clearAllAccounts();
+    }
+
+    @After
+    public void teardown() {
+        kinClient.clearAllAccounts();
+    }
+
+    @Test(expected = AccountDeletedException.class)
+    public void activateSync_DeletedAccount_AccountDeletedException() throws Exception {
+        KinAccount kinAccount = kinClient.addAccount();
+        kinClient.deleteAccount(0);
+        kinAccount.activateSync();
+    }
+
+    @Test(expected = AccountDeletedException.class)
+    public void getBalanceSync_DeletedAccount_AccountDeletedException() throws Exception {
+        KinAccount kinAccount = kinClient.addAccount();
+        kinClient.deleteAccount(0);
+        kinAccount.getBalanceSync();
+    }
+
+    @Test(expected = AccountDeletedException.class)
+    public void getStatusSync_DeletedAccount_AccountDeletedException() throws Exception {
+        KinAccount kinAccount = kinClient.addAccount();
+        kinClient.deleteAccount(0);
+        kinAccount.getStatusSync();
+    }
+
+    @Test(expected = AccountDeletedException.class)
+    public void sendTransactionSync_DeletedAccount_AccountDeletedException() throws Exception {
+        KinAccount kinAccount = kinClient.addAccount();
+        kinClient.deleteAccount(0);
+        kinAccount.sendTransactionSync("GBA2XHZRUAHEL4DZX7XNHR7HLBAUYPRNKLD2PIUKWV2LVVE6OJT4NDLM",
+            new BigDecimal(10));
+    }
+
+    @Test
+    public void getPublicAddress_DeletedAccount_EmptyPublicAddress() throws Exception {
+        KinAccount kinAccount = kinClient.addAccount();
+        kinClient.deleteAccount(0);
+        assertNull(kinAccount.getPublicAddress());
+    }
+
+}
